server/requirements.txt
python-telegram-bot==20.5
Django==4.2.5
psycopg2-binary==2.9.7
djangorestframework==3.14.0
